authorization is done

+++ backend/routes/posts.js
@@ -40,7 +40,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId,
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -70,9 +71,13 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+     if(result.modifiedCount > 0){
+       res.status(200).json({ message: "Update successful!" });
+     } else{
+      res.status(401).json({ message: "Not authorized!" });
+     }
+
     });
   }
 );
@@ -110,9 +115,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if(result.deleteOne  > 0){
+      res.status(200).json({ message: "Deleted successful!" });
+    } else{
+     res.status(401).json({ message: "Not authorized!" });
+    }
+
   });
 });
